denEngine/pyproject.toml
[project]
name = "denengine"
version = "0.1.0"
description = ""
authors = [
    {name = "Neel",email = "patelxneel@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "supabase (>=2.15.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.11.5"
